+++ test/test_helper.exs
@@ -1,4 +1,4 @@
 ExUnit.start()
-Fixate.start()
+Fixate.start(fixture_path: "test/fixtures")
 
 Fixate.add_parser("integer.txt", &String.to_integer(&1))